03-data-orchestration/requirements.txt
apache-airflow==2.8.0
apache-airflow-providers-snowflake==4.0.0
apache-airflow-providers-http==4.7.0
apache-airflow-providers-slack==7.2.0

# Database
snowflake-sqlalchemy==1.5.0

# Data processing
pandas==2.0.3
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code quality
black==23.7.0
pylint==2.17.4
flake8==6.0.0
isort==5.12.0
